requirements.txt
ipython==8.10.0
boto3==1.26.84
Django==4.1.3
django-extensions==3.2.1
django-oauth-toolkit==2.2.0
djangorestframework==3.14.0
grpcio==1.51.1
social-auth-app-django==5.0.0
social-auth-core==4.3.0
opensearch-py==2.1.1
protobuf==4.22.1
pydantic==1.10.2
psycopg2-binary==2.9.5
redis==4.4.2
requests==2.28.1
sentence-transformers==2.2.2
uwsgi==2.0.21
django_prometheus==2.2.0
drf-spectacular==0.25.1
PyYAML==6.0
git+https://github.com/ansible/ansible-risk-insight#egg=ansible_risk_insight==0.1.0
uwsgi-readiness-check==0.2.0
segment-analytics-python==2.2.2
pytz
django-health-check==3.17.0
